pyproject.toml
[tool.poetry]
name = "salesforecasttest"
version = "0.1.0"
description = ""
authors = ["m"]

[tool.poetry.dependencies]
python = ">=3.7.4,"
pandas = "^1.2.3"
Keras = "^2.4.3"
tensorflow-gpu = "^2.4.1"
matplotlib = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
